requirements.txt
# RealSense
pyrealsense2>=2.55.0

# Image & GUI
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.23.0,<1.24
pyyaml>=6.0
requests>=2.28.0

# TorchServe (로컬 실행)
torch==1.13.1
torchvision==0.14.1
torchserve==0.9.0
torch-model-archiver==0.9.0
torch-workflow-archiver==0.2.11

# OpenMMLab (포즈 추정용)
mmcv-full==1.7.0
mmdet==2.27.0
mmpose==0.29.0

# 유틸
scikit-image>=0.19.0
scipy>=1.9.0
